create issues

+++ src/db/index.ts
@@ -30,7 +30,7 @@ export const initDB = async () => {
         CHECK (type IN ('bug', 'feature_request')),
       status VARCHAR(20) NOT NULL DEFAULT 'open'
         CHECK (status IN ('open', 'in_progress', 'resolved')),
-      reporter_id BIGINT NOT NULL,
+      reporter_id INTEGER NOT NULL,
       created_at TIMESTAMP DEFAULT NOW(),
       updated_at TIMESTAMP DEFAULT NOW()
       )
